feat: Eliminar proveedores obsoletos y mover la clase QuizData a quiz_loader.dart
